requirements.txt
langchain
chromadb
streamlit
python-dotenv
GitPython
sentence-transformers
accelerate 
bitsandbytes